+++ storage_drivers/ontap/ontap_nas_flexgroup.go
@@ -421,6 +421,7 @@ func (d *NASFlexGroupStorageDriver) CreateFollowup(
 ) error {
 	volConfig.AccessInfo.NfsServerIP = d.Config.DataLIF
 	volConfig.AccessInfo.NfsPath = "/" + volConfig.InternalName
+	volConfig.AccessInfo.MountOptions = strings.TrimPrefix(d.Config.NfsMountOptions, "-o ")
 	volConfig.FileSystem = ""
 	return nil
 }